requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# Dépendances pour le pipeline ETL
pandas
psycopg2-binary
requests
sqlalchemy
loguru
pyyaml
python-dotenv
pytest
pytest-cov
